+++ veloq-runtime/src/tests/udp.rs
@@ -150,7 +150,7 @@ fn test_udp_echo() {
 
         // Verify
         assert_eq!(from_addr, server_addr);
-        assert_eq!(bytes_sent as u32, bytes_received);
+        assert_eq!(bytes_sent, bytes_received);
         assert_eq!(&recv_buf.as_slice()[..bytes_received as usize], test_data);
         println!("UDP echo test successful!");
